+++ crates/cubecl-wgpu/src/compute/mod.rs
@@ -3,8 +3,8 @@ mod storage;
 
 pub(super) mod mem_manager;
 pub(super) mod poll;
+pub(super) mod schedule;
 pub(super) mod stream;
-pub(super) mod tasks;
 pub(super) mod timings;
 
 mod server;
